+++ src/eg_futures.rs
@@ -1,3 +1,5 @@
+#![allow(dead_code)]
+
 use std::future::Future;
 use std::pin::Pin;
 use std::task::{Context, Poll};
