+++ src/main.rs
@@ -1,51 +1,12 @@
-use std::any::Any;
-use std::collections::HashMap;
-use std::future::Future;
-use std::pin::Pin;
-use std::time::Instant;
+use std::time::{Instant};
 use crate::eg_futures::{slow_future, TimerFuture};
-use crate::task_runner::Pool;
+use crate::executor::Executor;
 
 mod waker;
 mod eg_futures;
 mod task_runner;
-
-pub type Erased = Box<dyn Any + Send>;
-pub type BoxedFuture<T> = Pin<Box<dyn Future<Output = T> + Send>>;
-
-#[derive(Default)]
-pub struct IdGenerator {
-    next: usize
-}
-
-impl Iterator for IdGenerator {
-    type Item = Id;
-
-    fn next(&mut self) -> Option<Self::Item> {
-        if self.next == usize::MAX {
-            None
-        } else {
-            let ret = Some(Id {
-                index: self.next
-            });
-            self.next += 1;
-            ret
-        }
-    }
-}
-
-#[derive(Copy, Clone, Default, Hash, Debug, Eq, PartialEq)]
-pub struct Id {
-    index: usize,
-}
-
-impl Id {
-    pub fn next (&self) -> Self {
-        Self {
-            index: self.index + 1
-        }
-    }
-}
+mod ids;
+mod executor;
 
 #[macro_export]
 macro_rules! prt {
@@ -54,92 +15,6 @@ macro_rules! prt {
             println!($($arg)*)
         }
     }};
-}
-
-pub struct Running {
-    pool: Pool,
-    id_generator: IdGenerator
-}
-pub struct Finished;
-
-pub struct Executor<Stage> {
-    results_cache: HashMap<Id, Erased>,
-    stage_details: Stage,
-}
-
-impl Executor<Running> {
-    pub fn start() -> Self {
-        Executor {
-            results_cache: HashMap::new(),
-            stage_details: Running {
-                pool: Pool::new::<16>(),
-                id_generator: IdGenerator::default()
-            }
-        }
-    }
-
-    pub fn run<F: Future + Send + 'static> (&mut self, f: F) -> Option<Id>
-    where F::Output: Send
-    {
-        let id = self.stage_details.id_generator.next();
-        if let Some(id) = id {
-            self.stage_details.pool.run_future(id, Box::pin(async {
-                let res = f.await;
-                let erased: Erased = Box::new(res);
-                erased
-            }));
-        }
-        id
-    }
-
-    pub fn take_result<T: 'static> (&mut self, id: &Id) -> FutureResult<T> {
-        self.results_cache.extend(self.stage_details.pool.collect_results());
-        match self.results_cache.remove(id) {
-            None => FutureResult::NonExistent,
-            Some(x) => match x.downcast() {
-                Ok(t) => FutureResult::Expected(*t),
-                Err(b) => FutureResult::Other(b),
-            }
-        }
-    }
-
-    pub fn join (mut self) -> Executor<Finished> {
-        self.results_cache.extend(self.stage_details.pool.join());
-
-        Executor {
-            results_cache: self.results_cache,
-            stage_details: Finished
-        }
-    }
-}
-
-impl Executor<Finished> {
-    pub fn take_result<T: 'static> (&mut self, id: &Id) -> FutureResult<T> {
-        match self.results_cache.remove(id) {
-            None => FutureResult::NonExistent,
-            Some(x) => match x.downcast() {
-                Ok(t) => FutureResult::Expected(*t),
-                Err(b) => FutureResult::Other(b),
-            }
-        }
-    }
-}
-
-#[derive(Debug)]
-pub enum FutureResult<T> {
-    Expected(T),
-    Other(Box<dyn Any>),
-    NonExistent
-}
-
-impl<T> FutureResult<T> {
-    pub fn unwrap (self) -> T {
-        match self {
-            Self::Expected(e) => e,
-            Self::Other(_) => panic!("failed to unwrap FutureResult as found wrong type"),
-            Self::NonExistent => panic!("failed to unwrap FutureResult as `None`")
-        }
-    }
 }
 
 
@@ -160,14 +35,16 @@ fn main() {
     };
 
     let start = Instant::now();
-    let id1 = executor.run(create_task(150, 1)).unwrap();
-    let id2 = executor.run(create_task(150, 2)).unwrap();
-    let id3 = executor.run(create_task(50, 3)).unwrap();
-    let id4 = executor.run(create_task(200, 4)).unwrap();
+    let instant = executor.run(create_task(0, 0)).unwrap();
+    let id1 = executor.run(create_task(15, 1)).unwrap();
+    let id2 = executor.run(create_task(5, 2)).unwrap();
+    let id3 = executor.run(create_task(15, 3)).unwrap();
+    let id4 = executor.run(create_task(20, 4)).unwrap();
     
-    let fib = executor.run(slow_future(175)).unwrap();
+    let fib = executor.run(slow_future(18)).unwrap();
     
-    // let fail: u32 = executor.take_result(&id4).unwrap();
+    let instant: Option<u64> = executor.take_result(&instant).into();
+    println!("[main] instant result actually instant? {instant:?}");
 
     println!("[main] created all tasks, joining executor");
     let mut executor = executor.join();
@@ -180,6 +57,10 @@ fn main() {
     let res4: u64 = executor.take_result(&id4).unwrap();
     let fib: u64 = executor.take_result(&fib).unwrap();
 
-    println!("[main] got results ({res1:?},{res2:?},{res3:?},{res4:?})");
+    assert_eq!(res1, 15);
+    assert_eq!(res2, 5);
+    assert_eq!(res3, 15);
+    assert_eq!(res4, 20);
+
     println!("[main] slow calc is {fib:?}");
 }
